requirements.txt
Flask==3.0.0
PyYAML==6.0.1
pymysql==1.1.0
clickhouse-connect==0.6.23
requests==2.31.0

